Use int64_t type for shape dims

+++ chainerx_cc/chainerx/native/native_device/linalg.cc
@@ -191,9 +191,9 @@ std::tuple<Array, Array> QRImpl(const Array& a, QRMode mode) {
     Device& device = a.device();
     Dtype dtype = a.dtype();
 
-    int m = a.shape()[0];
-    int n = a.shape()[1];
-    int mn = std::min(m, n);
+    int64_t m = a.shape()[0];
+    int64_t n = a.shape()[1];
+    int64_t mn = std::min(m, n);
 
     Array Q = Empty(Shape({0}), dtype, device);
     Array R = a.Transpose().Copy();  // QR decomposition is done in-place
@@ -227,7 +227,7 @@ std::tuple<Array, Array> QRImpl(const Array& a, QRMode mode) {
         return std::make_tuple(std::move(R), std::move(tau));
     }
 
-    int mc;
+    int64_t mc;
     if (mode == QRMode::complete && m > n) {
         mc = m;
         Q = Empty(Shape({m, m}), dtype, device);
